Refactor io -> readers
