+++ sarracen/__init__.py
@@ -1,7 +1,7 @@
 
-from sarracen.io.read_marisa import read_marisa
-from sarracen.io.read_csv import read_csv
-from sarracen.io.read_phantom import read_phantom
+from sarracen.readers.read_marisa import read_marisa
+from sarracen.readers.read_csv import read_csv
+from sarracen.readers.read_phantom import read_phantom
 
 from sarracen.sarracen_dataframe import SarracenDataFrame
